backend/github-repo-leaderboard/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
PyGithub
scikit-learn
numpy
google-generativeai
